requirements.txt
requests==2.32.3
PyYAML==6.0.2
google-cloud-bigquery==3.25.0
google-cloud-carbon-footprint==0.1.0
google-auth==2.34.0
